NOTICKET Add app folder. Move files and change paths. Remove app set vars and set to config.
