+++ app/lib/LAUNCH_MESSAGE.txt
