+++ app/lib/app.js
@@ -11,25 +11,26 @@
 
 // Module dependencies.
 var bodyParser = require('body-parser');
-var config = require('../config');
 var cookieParser = require('cookie-parser');
-var db = require('./db');
 var exphbs = require('express-handlebars');
 var express = require('express');
 var favicon = require('serve-favicon');
 var flash = require('connect-flash');
 var fs = require('fs');
-var logging = require('./logging');
 var methodOverride = require('method-override');
-var middleware = require('./middleware');
 var mongoose = require('mongoose');
 var morgan = require('morgan');
-var packageJson = require('../package.json');
 var passport = require('passport');
 var path = require('path');
-var router = require('./router');
 var session = require('express-session');
 var winston = require('winston');
+
+// App Modules
+var config = require('../../config');
+var db = require('./db');
+var logging = require('./logging');
+var middleware = require('./middleware');
+var router = require('./router');
 
 // Controllers
 var ErrorController = require('../controllers/errorController');
@@ -39,33 +40,36 @@ var OAuthController = require('../controllers/oAuthController');
 
 global.App = {
     app: express(),
-    port: process.env.PORT,
-    version: packageJson.version,
-    root: path.join(__dirname, '..'),
-    env: process.env.port,
+    appPath: path.join(__dirname, '..'),
+    rootPath: path.join(__dirname, '..', '..'),
 
 
+    /**
+     * Initialise the app. Pull in all middlewares, routing, etc.
+     */
     init: function() {
         logging.init();  // Logging through Winston
         OAuthController.initOauth();  // Authentication
         db.connect();  // Mongo database
 
 
-        this.app.engine('hbs', exphbs({defaultLayout: 'main', extname: 'hbs'}));  // Templating
+        // Templating
+        this.app.engine('hbs', exphbs({
+            defaultLayout: 'main',
+            extname: 'hbs',
+            layoutsDir: path.join(this.appPath, 'views', 'layouts'),
+            partialsDir: path.join(this.appPath, 'views', 'partials')
+        }));
         this.app.set('view engine', 'hbs');
-        // TODO(allard): Move to config. Add default.
-        // TODO(allard): Should I be doing this here or be putting all of this in config and ignoring app.set?
-        this.app.set('port', config.http.PORT);
-        this.app.set('env', config.http.ENV);
-        this.app.set('package', packageJson);
+        this.app.set('views', path.join(this.appPath, 'views'));
 
         // Pre-routing Middleware.
         this.app.use(bodyParser.urlencoded({extended: false}));
         this.app.use(methodOverride(middleware.parseRequestMethodValue));
         this.app.use(morgan('dev'));  // Format logging with Morgan
         // Favicon and Static paths need to go before Session middleware to avoid superfluous session creation.
-        this.app.use(favicon(path.join('.', 'public', 'favicon.ico')));
-        this.app.use(express.static(path.join('.', 'public')));
+        this.app.use(favicon(path.join(this.rootPath, 'public', 'favicon.ico')));
+        this.app.use(express.static(path.join(this.rootPath, 'public')));
         // Session Cookie Middleware
         this.app.use(cookieParser());
         this.app.use(session({
@@ -80,6 +84,7 @@ global.App = {
         this.app.use(flash());
         this.app.use(passport.initialize());
         this.app.use(passport.session());
+        // TODO(allard): Move this to the session serialisation function.
         this.app.use(UserController.createOrUpdate);  // Save the user to the DB
         this.app.use(middleware.putConfigInLocals);
 
@@ -91,25 +96,31 @@ global.App = {
         this.app.use(ErrorController.renderError);
     },
 
+    /**
+     * Start the server to listen on the configured port.
+     */
     start: function() {
+        winston.info('Starting the app...');
+
         if (!this.started) {
             this.started = true;
-            // Run the server.
-            // TODO(allard): Bind isn't working here... Don't know what I'm doing wrong...
-            this.app.listen(this.app.get('port'), logSuccessfulLaunch);
+            this.app.listen(config.http.PORT, logSuccessfulLaunch);
         }
     }
 };
 
 
+/**
+ * Log a succesful launch.
+ */
 var logSuccessfulLaunch = function() {
-    fs.readFile(path.join(__dirname, '..', 'fixtures', 'LAUNCH_MESSAGE.txt'), 'utf-8', function(err, data) {
+    fs.readFile(path.join(App.appPath, 'lib', 'LAUNCH_MESSAGE.txt'), 'utf-8', function(err, launchMessage) {
         if (err) {
             return winston.error('Read file error:', err);
         }
-        winston.warn(data);
-        winston.warn('App running in %s mode.', App.app.get('env'));
-        winston.warn('Server listening on http://localhost:%d.', App.app.get('port'));
+        winston.warn(launchMessage);
+        winston.info('App running in %s mode.', config.http.ENV);
+        winston.info('Server listening on http://localhost:%d.', config.http.PORT);
     });
 };
